Created a product folder in the components directory, and moved components into it.
