+++ app/components/product/ProductDetails.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import prisma from '../lib/prisma';
+import prisma from '../../lib/prisma';
 import Image from 'next/image';
 import NoProductFound from './NoProductFound';
